removed the loop from the render function and added an argument to it. Added a local storage so that books stay on the page when added even after a refresh or tab close
