+++ src/index.js
@@ -1,7 +1,8 @@
 import { DomManipulation } from "/src/domManipulation.js"
-import { addBookToLibrary } from "/src/addingBooks.js"
-import { render } from "/src/renderBook.js"
+import { addBookToLibrary, render, createBook } from "/src/addingBooks.js"
+//import { render } from "/src/renderBook.js"
 import {myLibrary} from "/src/addingBooks.js"
+
 
 const domManip = new DomManipulation()
 domManip.openModal();
@@ -10,7 +11,8 @@ const addBookButton = document.querySelector('.add-book')
 addBookButton.addEventListener('click', () => {
 
     addBookToLibrary()
-    render()
+    // render()
+    createBook();
     console.log(myLibrary)
 });
